updated IO to use generics.
