+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/TimothyStiles/poly
 
-go 1.16
+go 1.18
 
 require (
 	github.com/google/go-cmp v0.5.6
@@ -8,8 +8,11 @@ require (
 	github.com/mroth/weightedrand v0.4.1
 	github.com/pmezard/go-difflib v1.0.0
 	github.com/sergi/go-diff v1.2.0
-	github.com/urfave/cli/v2 v2.3.0
+	lukechampine.com/blake3 v1.1.5
+)
+
+require (
+	github.com/klauspost/cpuid v1.3.1 // indirect
 	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
-	lukechampine.com/blake3 v1.1.5
 )
